ISSD-1342 Add Types Folder
